+++ psutil/_pswindows.py
@@ -368,8 +368,7 @@ def net_if_addrs():
 def sensors_battery():
     """Return battery information."""
     # For constants meaning see:
-    # https://msdn.microsoft.com/en-us/library/windows/desktop/
-    #     aa373232(v=vs.85).aspx
+    # https://msdn.microsoft.com/en-us/library/windows/desktop/aa373232(v=vs.85).aspx
     acline_status, flags, percent, secsleft = cext.sensors_battery()
     power_plugged = acline_status == 1
     no_battery = bool(flags & 128)
@@ -794,21 +793,36 @@ class Process:
         return ntp.pmem(
             rss=d["WorkingSetSize"],
             vms=d["PrivateUsage"],
-            num_page_faults=d["PageFaultCount"],
-            paged_pool=d["QuotaPagedPoolUsage"],
-            nonpaged_pool=d["QuotaNonPagedPoolUsage"],
             peak_rss=d["PeakWorkingSetSize"],
             peak_vms=d["PeakPagefileUsage"],
-            peak_paged_pool=d["QuotaPeakPagedPoolUsage"],
-            peak_nonpaged_pool=d["QuotaPeakNonPagedPoolUsage"],
+            _deprecated={
+                # old aliases
+                "wset": d["WorkingSetSize"],  # 'rss'
+                "peak_wset": d["PeakWorkingSetSize"],  # 'peak_rss'
+                "pagefile": d["PrivateUsage"],  # 'vms'
+                "private": d["PrivateUsage"],  # 'vms'
+                "peak_pagefile": d["PeakPagefileUsage"],  # 'vms'
+                # fields which were moved to memory_info_ex()
+                "paged_pool": d["QuotaPagedPoolUsage"],
+                "nonpaged_pool": d["QuotaNonPagedPoolUsage"],
+                "peak_paged_pool": d["QuotaPeakPagedPoolUsage"],
+                "peak_nonpaged_pool": d["QuotaPeakNonPagedPoolUsage"],
+                # moved to page_faults()
+                "num_page_faults": d["PageFaultCount"],
+            },
         )
 
     @wrap_exceptions
     def memory_info_ex(self):
         d = self._oneshot()
+        raw = self._get_raw_meminfo()
         return {
             "virtual": d["VirtualSize"],
             "peak_virtual": d["PeakVirtualSize"],
+            "paged_pool": raw["QuotaPagedPoolUsage"],
+            "nonpaged_pool": raw["QuotaNonPagedPoolUsage"],
+            "peak_paged_pool": raw["QuotaPeakPagedPoolUsage"],
+            "peak_nonpaged_pool": raw["QuotaPeakNonPagedPoolUsage"],
         }
 
     @wrap_exceptions
@@ -819,8 +833,8 @@ class Process:
 
     @wrap_exceptions
     def page_faults(self):
-        ret = cext.proc_page_faults(self.pid)
-        return ntp.ppagefaults(*ret)
+        t = cext.proc_page_faults(self.pid)
+        return ntp.ppagefaults(*t)
 
     def memory_maps(self):
         try:
@@ -877,9 +891,7 @@ class Process:
             # We'll just rely on the internal polling and return None
             # when the PID disappears. Subprocess module does the same
             # (return None):
-            # https://github.com/python/cpython/blob/
-            #     be50a7b627d0aa37e08fa8e2d5568891f19903ce/
-            #     Lib/subprocess.py#L1193-L1194
+            # https://github.com/python/cpython/blob/be50a7b627d0/Lib/subprocess.py#L1193-L1194
             exit_code = None
 
         # At this point WaitForSingleObject() returned WAIT_OBJECT_0,